REFACTOR: users controllers and improve hexagonal architecture
