+++ user_service/src/domain/datasources/auth.datasource.ts
@@ -1,0 +1,6 @@
+import { UserEntity } from '../entities/user.entity';
+
+export abstract class AuthDatasource {
+  abstract Create(user: UserEntity): Promise<UserEntity | void>;
+  abstract UpdateVerifyUser(userId: number): Promise<UserEntity | null>;
+}
