+++ user_service/src/domain/repositories/auth.repository.ts
@@ -1,0 +1,6 @@
+import { UserEntity } from '../entities/user.entity';
+
+export abstract class AuthRepository {
+  abstract Create(user: UserEntity): Promise<UserEntity | void>;
+  abstract UpdateVerifyUser(userId: number): Promise<UserEntity | null>;
+}
