+++ user_service/src/domain/repositories/index.ts
@@ -1,0 +1,2 @@
+export { AuthRepository } from './auth.repository';
+export { UserRepository } from './user.repository';
