+++ user_service/src/infrastructure/datasources/index.ts
@@ -1,0 +1,2 @@
+export { PgAuthDatasource } from './auth-pg.datasource';
+export { PgUserDatasource } from './user-pg.datasource';
