+++ user_service/src/infrastructure/repositories/index.ts
@@ -1,0 +1,2 @@
+export { AuthRepositoryImpl } from './auth.repository';
+export { UserRepositoryImpl } from './user.repository';
