+++ user_service/src/infrastructure/services/email.service.ts
@@ -1,4 +1,4 @@
-import nodemailer, { Transporter, createTransport } from 'nodemailer';
+import { Transporter, createTransport } from 'nodemailer';
 
 export interface SendMailOptions {
   to: string | string[];
@@ -25,40 +25,17 @@ export class EmailService {
     const { to, subject, html, attachments = [] } = options;
 
     try {
-      const sentInformation = (await this.transporter.sendMail({
+      (await this.transporter.sendMail({
         to,
         subject,
         html,
         attachments,
       })) as SendMailOptions;
 
-      console.log(sentInformation);
       return true;
     } catch (error) {
       console.log(error);
       return false;
     }
-  }
-
-  async sendEmailWithFileSystemLogs(to: string | string[]) {
-    const subject = 'Logs del servidor';
-    const html = `
-    <h3>Logs de sistema - NOC</h3>
-    <p>Lorem velit non veniam ullamco ex eu laborum deserunt est amet elit nostrud sit. Dolore ullamco duis in ut deserunt. Ad pariatur labore exercitation adipisicing excepteur elit anim eu consectetur excepteur est dolor qui. Voluptate consectetur proident ex fugiat reprehenderit exercitation laboris amet Lorem ullamco sit. Id aute ad do laborum officia labore proident laborum. Amet sit aliqua esse anim fugiat ut eu excepteur veniam incididunt occaecat sit irure aliquip. Laborum esse cupidatat adipisicing non et cupidatat ut esse voluptate aute aliqua pariatur.</p>
-    <p>Ver logs adjuntos</p>
-    `;
-
-    const attachments: Attachement[] = [
-      { filename: 'logs-all.log', path: './logs/logs-all.log' },
-      { filename: 'logs-high.log', path: './logs/logs-high.log' },
-      { filename: 'logs-medium.log', path: './logs/logs-medium.log' },
-    ];
-
-    return this.sendEmail({
-      to,
-      subject,
-      attachments,
-      html,
-    });
   }
 }
