+++ user_service/src/infrastructure/use-cases/users/index.ts
@@ -1,0 +1,2 @@
+export { GetUserUseCase } from './get-user';
+export { GetUsersUseCase } from './get-users';
