+++ user_service/src/user.api.ts
@@ -3,11 +3,10 @@ import { APIGatewayProxyEvent } from 'aws-lambda';
 
 import { PgUserDatasource } from './infrastructure/datasources/user-pg.datasource';
 import { UserRepositoryImpl } from './infrastructure/repositories/user.repository';
-import { EmailService } from './infrastructure/services/email.service';
-import { UserService } from './presentation/services/user.service';
+import { UserController } from './presentation/controllers/user.controller';
 import { ErrorResponse } from './presentation/utils';
 
-const service = new UserService(new UserRepositoryImpl(new PgUserDatasource()), new EmailService());
+const service = new UserController(new UserRepositoryImpl(new PgUserDatasource()));
 
 export const handler = (event: APIGatewayProxyEvent) => {
   const isRoot = event.pathParameters === null;
